首卡常开启用后，才需要删除取消首卡常卡权限的人员

+++ src/main/java/com/xiangshangban/device/service/impl/EntranceGuardServiceImpl.java
@@ -784,7 +784,10 @@ public class EntranceGuardServiceImpl implements IEntranceGuardService {
             doorSettingMapper.updateByPrimaryKeySelective(doorSetting);
         }
         //删除不在新列表中的人员权限
-        int deleteNum = timeRangePrivilegeEmployeeMapper.deletePrivilege(doorId,employeeIdList);
+        if("1".equals(enableFirstCardKeepOpen)){
+        	int deleteNum = timeRangePrivilegeEmployeeMapper.deletePrivilege(doorId,employeeIdList);
+        }
+        
         //遍历人员
         for (String employeeId : employeeIdList) {
